fix: allow 0 (unlimited) filesystem_watchers.max_events which is the new default, except for windows at 1000 (#3279)

+++ lua/nvim-tree/explorer/watch.lua
@@ -79,12 +79,12 @@ function M.create_watcher(node)
     watcher.data.outstanding_events = watcher.data.outstanding_events + 1
 
     -- disable watcher when outstanding exceeds max
-    if watcher.data.outstanding_events > M.config.filesystem_watchers.max_events then
+    if M.config.filesystem_watchers.max_events > 0 and watcher.data.outstanding_events > M.config.filesystem_watchers.max_events then
       notify.error(string.format(
         "Observed %d consecutive file system events with interval < %dms, exceeding filesystem_watchers.max_events=%s. Disabling watcher for directory '%s'. Consider adding this directory to filesystem_watchers.ignore_dirs",
         watcher.data.outstanding_events,
-        M.config.filesystem_watchers.max_events,
         M.config.filesystem_watchers.debounce_delay,
+        M.config.filesystem_watchers.max_events,
         node.absolute_path
       ))
       node:destroy_watcher()
